này code thêm cái quản lí nghỉ phép

+++ src/main/java/vn/thangloi/spring/hrm/service/PositionServiceImpl.java
@@ -3,9 +3,12 @@ package vn.thangloi.spring.hrm.service;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import vn.thangloi.spring.hrm.dao.PositionRepository;
+import vn.thangloi.spring.hrm.entity.Department;
+import vn.thangloi.spring.hrm.entity.Employee;
 import vn.thangloi.spring.hrm.entity.Position;
 
 import java.util.List;
+import java.util.Set;
 
 @Service
 public class PositionServiceImpl implements PositionService {
@@ -39,7 +42,17 @@ public class PositionServiceImpl implements PositionService {
 
     @Override
     public void deletePositionById(int id) {
-        positionRepository.deleteById(id);
+        Position position = positionRepository.findById(id).orElse(null);
+        if (position != null) {
+            Set<Employee> employees = position.getEmployees();
+            if (!employees.isEmpty()) {
+                throw new RuntimeException("Không thể xóa vị trí này vì có nhân viên đang thuộc vị trí này.");
+            } else {
+                positionRepository.deleteById(id);
+            }
+        } else {
+            throw new RuntimeException("Không có id vị trị này.");
+        }
     }
